requirements.txt
PySide6
PyYAML
setuptools